services/web/requirements.txt
Flask==1.1.2
gunicorn==20.1.0
redis==3.5.3
python-redis-lock==3.7.0
